1. keeping track on insurance Company Project-250410-1740

+++ src/_250407_Insurance_Company_Project/Insurance/Insurance.java
@@ -43,6 +43,6 @@ public abstract class Insurance {
         this.endDate = endDate;
     }
 
-    abstract int calculate();
+    abstract double calculate();
 
 }
